Refactor monitoring alerts
